When policies.namespaceSupport=permissive, AstNamespace now behaves like a nested AstPackage (the common functionality is moved to an abstract base class AstModule)
