+++ libraries/api-extractor/src/extractor/api-extractor.schema.json
@@ -74,7 +74,6 @@
           "enum": [ "conservative", "permissive" ]
         }
       },
-      "required": [ ],
       "additionalProperties": false
     },
